pyproject.toml
[project]
name = "ai-customer-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.63",
    "faiss-cpu>=1.13.2",
    "langfuse>=3.14.5",
    "mcp[http]>=1.26.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "strands-agents[ollama,openai]>=1.29.0",
]
